+++ src/Allegro.Extensions.Cqrs/Allegro.Extensions.Cqrs.Abstractions/Commands/ICommand.cs
@@ -1,8 +1,0 @@
-namespace Allegro.Extensions.Cqrs.Abstractions.Commands;
-
-/// <summary>
-/// Cqrs command marker
-/// </summary>
-public interface ICommand
-{
-}