+++ src/Allegro.Extensions.Cqrs/Allegro.Extensions.Cqrs.FluentValidations/FluentQueryValidator.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.DependencyInjection;
 namespace Allegro.Extensions.Cqrs.FluentValidations;
 
 internal class FluentQueryValidator<T> : IQueryValidator<T>
-    where T : IQuery
+    where T : Query
 {
     private readonly IServiceProvider _serviceProvider;